techpack/audio: Don't build HDMI codec if external display driver isn't built
HDMI codec driver, as on ASoC counterpart on older MSM kernels, highly depends
on external display (HDMI on older kernels) driver, and shouldn't be built if
the main dependency isn't built.

  ERROR: "msm_ext_disp_register_audio_codec" [techpack/audio/asoc/codecs/hdmi_dlkm.ko] undefined!
  make[2]: *** [/mnt/ssd/kernels/msm-4.9/scripts/Makefile.modpost:107: __modpost] Error 1
  make[1]: *** [/mnt/ssd/kernels/msm-4.9/Makefile:1354: modules] Error 2
  make[1]: *** Waiting for unfinished jobs....

Removing the driver config from msm8909auto config is stupid, as it means that
QM215 will lose it as well.

Change-Id: 7e1f420a1c9a4ef9a9f0ba637ef90243425e78e1
Signed-off-by: Albert I <kras@raphielgang.org>
Signed-off-by: TogoFire <italomellopereira@gmail.com>